removed memory leaks, validated with adress sanitizer
Things to do after this:
1. create copy/move and copy/move constructors for map
2. move from shared_ptr to unique_ptr for Cont
3. create correct contructors for Cont
4. check with other sanitizers
