+++ CMakeLists.txt
@@ -1,33 +1,34 @@
 cmake_minimum_required(VERSION 3.7...3.19)
 
-if(${CMAKE_VERSION} VERSION_LESS 3.12)
+if (${CMAKE_VERSION} VERSION_LESS 3.12)
     cmake_policy(VERSION ${CMAKE_MAJOR_VERSION}.${CMAKE_MINOR_VERSION})
-endif()
+endif ()
 
 project(PO2 VERSION 1.0
-                  DESCRIPTION "project ontwerp 2"
-                  LANGUAGES CXX)
+        DESCRIPTION "project ontwerp 2"
+        LANGUAGES CXX)
 
 set(CMAKE_CXX_STANDARD 17)
-set(CMAKE_CXX_FLAGS "-o3  -march=native")
+# add "-fsanitize=address -fno-omit-frame-pointer" to detect memory errors. remove after use, since it slows down program significantly
+set(CMAKE_CXX_FLAGS "-o3  -march=native -fsanitize=address -fno-omit-frame-pointer")
 set(CMAKE_EXPORT_COMPILE_COMMANDS ON)
 
 set(CLANG_TIDY_CXX -checks=-*,
-    bugprone-*,
-    cert-*,
-    clang-analyzer-*,
-    concurrency-*,
-    cppcoreguidelines-*,
-    hicpp-*,
-    llvm-*,
-    misc-*,
-    modernize-*,
-    mpi-*,
-    openmp-*,
-    performance-*,
-    portability-*,
-    readability-*,
-)
+        bugprone-*,
+        cert-*,
+        clang-analyzer-*,
+        concurrency-*,
+        cppcoreguidelines-*,
+        hicpp-*,
+        llvm-*,
+        misc-*,
+        modernize-*,
+        mpi-*,
+        openmp-*,
+        performance-*,
+        portability-*,
+        readability-*,
+        )
 set(CLANG_TIDY_SUPPRESS_CXX
     misc-*,
 )
