--- cmake-build-debug/CMakeFiles/LP-test.dir/build.make
@@ -1,118 +1,0 @@
-# CMAKE generated file: DO NOT EDIT!
-# Generated by "Unix Makefiles" Generator, CMake Version 3.17
-
-# Delete rule output on recipe failure.
-.DELETE_ON_ERROR:
-
-
-#=============================================================================
-# Special targets provided by cmake.
-
-# Disable implicit rules so canonical targets will work.
-.SUFFIXES:
-
-
-# Disable VCS-based implicit rules.
-% : %,v
-
-
-# Disable VCS-based implicit rules.
-% : RCS/%
-
-
-# Disable VCS-based implicit rules.
-% : RCS/%,v
-
-
-# Disable VCS-based implicit rules.
-% : SCCS/s.%
-
-
-# Disable VCS-based implicit rules.
-% : s.%
-
-
-.SUFFIXES: .hpux_make_needs_suffix_list
-
-
-# Command-line flag to silence nested $(MAKE).
-$(VERBOSE)MAKESILENT = -s
-
-# Suppress display of executed commands.
-$(VERBOSE).SILENT:
-
-
-# A target that is always out of date.
-cmake_force:
-
-.PHONY : cmake_force
-
-#=============================================================================
-# Set environment variables for the build.
-
-# The shell in which to execute make rules.
-SHELL = /bin/sh
-
-# The CMake executable.
-CMAKE_COMMAND = /var/lib/snapd/snap/clion/139/bin/cmake/linux/bin/cmake
-
-# The command to remove a file.
-RM = /var/lib/snapd/snap/clion/139/bin/cmake/linux/bin/cmake -E rm -f
-
-# Escaping for special characters.
-EQUALS = =
-
-# The top-level source directory on which CMake was run.
-CMAKE_SOURCE_DIR = /home/MassiveAtoms/Desktop/PO2
-
-# The top-level build directory on which CMake was run.
-CMAKE_BINARY_DIR = /home/MassiveAtoms/Desktop/PO2/cmake-build-debug
-
-# Include any dependencies generated for this target.
-include CMakeFiles/LP-test.dir/depend.make
-
-# Include the progress variables for this target.
-include CMakeFiles/LP-test.dir/progress.make
-
-# Include the compile flags for this target's objects.
-include CMakeFiles/LP-test.dir/flags.make
-
-CMakeFiles/LP-test.dir/test/hashmap_test.cpp.o: CMakeFiles/LP-test.dir/flags.make
-CMakeFiles/LP-test.dir/test/hashmap_test.cpp.o: ../test/hashmap_test.cpp
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green --progress-dir=/home/MassiveAtoms/Desktop/PO2/cmake-build-debug/CMakeFiles --progress-num=$(CMAKE_PROGRESS_1) "Building CXX object CMakeFiles/LP-test.dir/test/hashmap_test.cpp.o"
-	/usr/bin/g++  $(CXX_DEFINES) $(CXX_INCLUDES) $(CXX_FLAGS) -o CMakeFiles/LP-test.dir/test/hashmap_test.cpp.o -c /home/MassiveAtoms/Desktop/PO2/test/hashmap_test.cpp
-
-CMakeFiles/LP-test.dir/test/hashmap_test.cpp.i: cmake_force
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green "Preprocessing CXX source to CMakeFiles/LP-test.dir/test/hashmap_test.cpp.i"
-	/usr/bin/g++ $(CXX_DEFINES) $(CXX_INCLUDES) $(CXX_FLAGS) -E /home/MassiveAtoms/Desktop/PO2/test/hashmap_test.cpp > CMakeFiles/LP-test.dir/test/hashmap_test.cpp.i
-
-CMakeFiles/LP-test.dir/test/hashmap_test.cpp.s: cmake_force
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green "Compiling CXX source to assembly CMakeFiles/LP-test.dir/test/hashmap_test.cpp.s"
-	/usr/bin/g++ $(CXX_DEFINES) $(CXX_INCLUDES) $(CXX_FLAGS) -S /home/MassiveAtoms/Desktop/PO2/test/hashmap_test.cpp -o CMakeFiles/LP-test.dir/test/hashmap_test.cpp.s
-
-# Object files for target LP-test
-LP__test_OBJECTS = \
-"CMakeFiles/LP-test.dir/test/hashmap_test.cpp.o"
-
-# External object files for target LP-test
-LP__test_EXTERNAL_OBJECTS =
-
-LP-test: CMakeFiles/LP-test.dir/test/hashmap_test.cpp.o
-LP-test: CMakeFiles/LP-test.dir/build.make
-LP-test: CMakeFiles/LP-test.dir/link.txt
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green --bold --progress-dir=/home/MassiveAtoms/Desktop/PO2/cmake-build-debug/CMakeFiles --progress-num=$(CMAKE_PROGRESS_2) "Linking CXX executable LP-test"
-	$(CMAKE_COMMAND) -E cmake_link_script CMakeFiles/LP-test.dir/link.txt --verbose=$(VERBOSE)
-
-# Rule to build all files generated by this target.
-CMakeFiles/LP-test.dir/build: LP-test
-
-.PHONY : CMakeFiles/LP-test.dir/build
-
-CMakeFiles/LP-test.dir/clean:
-	$(CMAKE_COMMAND) -P CMakeFiles/LP-test.dir/cmake_clean.cmake
-.PHONY : CMakeFiles/LP-test.dir/clean
-
-CMakeFiles/LP-test.dir/depend:
-	cd /home/MassiveAtoms/Desktop/PO2/cmake-build-debug && $(CMAKE_COMMAND) -E cmake_depends "Unix Makefiles" /home/MassiveAtoms/Desktop/PO2 /home/MassiveAtoms/Desktop/PO2 /home/MassiveAtoms/Desktop/PO2/cmake-build-debug /home/MassiveAtoms/Desktop/PO2/cmake-build-debug /home/MassiveAtoms/Desktop/PO2/cmake-build-debug/CMakeFiles/LP-test.dir/DependInfo.cmake --color=$(COLOR)
-.PHONY : CMakeFiles/LP-test.dir/depend
-
