--- cmake-build-debug/CMakeFiles/LP-test.dir/cmake_clean.cmake
@@ -1,10 +1,0 @@
-file(REMOVE_RECURSE
-  "CMakeFiles/LP-test.dir/test/hashmap_test.cpp.o"
-  "LP-test"
-  "LP-test.pdb"
-)
-
-# Per-language clean rules from dependency scanning.
-foreach(lang CXX)
-  include(CMakeFiles/LP-test.dir/cmake_clean_${lang}.cmake OPTIONAL)
-endforeach()
