--- cmake-build-debug/CMakeFiles/LP-test.dir/flags.make
@@ -1,10 +1,0 @@
-# CMAKE generated file: DO NOT EDIT!
-# Generated by "Unix Makefiles" Generator, CMake Version 3.17
-
-# compile CXX with /usr/bin/g++
-CXX_FLAGS = -o3  -march=native -g   -std=gnu++17
-
-CXX_DEFINES = 
-
-CXX_INCLUDES = 
-
