--- cmake-build-debug/CMakeFiles/LP-test.dir/progress.make
@@ -1,3 +1,0 @@
-CMAKE_PROGRESS_1 = 1
-CMAKE_PROGRESS_2 = 2
-
