+++ cmake-build-debug/CMakeFiles/Makefile.cmake
@@ -21,9 +21,11 @@ set(CMAKE_MAKEFILE_DEPENDS
   "/var/lib/snapd/snap/clion/139/bin/cmake/linux/share/cmake-3.17/Modules/CMakeSystemSpecificInformation.cmake"
   "/var/lib/snapd/snap/clion/139/bin/cmake/linux/share/cmake-3.17/Modules/CMakeSystemSpecificInitialize.cmake"
   "/var/lib/snapd/snap/clion/139/bin/cmake/linux/share/cmake-3.17/Modules/Compiler/CMakeCommonCompilerMacros.cmake"
-  "/var/lib/snapd/snap/clion/139/bin/cmake/linux/share/cmake-3.17/Modules/Compiler/GNU-CXX.cmake"
+  "/var/lib/snapd/snap/clion/139/bin/cmake/linux/share/cmake-3.17/Modules/Compiler/Clang-CXX.cmake"
+  "/var/lib/snapd/snap/clion/139/bin/cmake/linux/share/cmake-3.17/Modules/Compiler/Clang.cmake"
   "/var/lib/snapd/snap/clion/139/bin/cmake/linux/share/cmake-3.17/Modules/Compiler/GNU.cmake"
   "/var/lib/snapd/snap/clion/139/bin/cmake/linux/share/cmake-3.17/Modules/Internal/CMakeCheckCompilerFlag.cmake"
+  "/var/lib/snapd/snap/clion/139/bin/cmake/linux/share/cmake-3.17/Modules/Platform/Linux-Clang-CXX.cmake"
   "/var/lib/snapd/snap/clion/139/bin/cmake/linux/share/cmake-3.17/Modules/Platform/Linux-GNU-CXX.cmake"
   "/var/lib/snapd/snap/clion/139/bin/cmake/linux/share/cmake-3.17/Modules/Platform/Linux-GNU.cmake"
   "/var/lib/snapd/snap/clion/139/bin/cmake/linux/share/cmake-3.17/Modules/Platform/Linux.cmake"
