+++ cmake-build-debug/CMakeFiles/benchmarks.dir/DependInfo.cmake
@@ -7,7 +7,7 @@ set(CMAKE_DEPENDS_CHECK_CXX
   "/home/MassiveAtoms/Desktop/PO2/benchmarks/bench.cpp" "/home/MassiveAtoms/Desktop/PO2/cmake-build-debug/CMakeFiles/benchmarks.dir/benchmarks/bench.cpp.o"
   "/home/MassiveAtoms/Desktop/PO2/benchmarks/generator.cpp" "/home/MassiveAtoms/Desktop/PO2/cmake-build-debug/CMakeFiles/benchmarks.dir/benchmarks/generator.cpp.o"
   )
-set(CMAKE_CXX_COMPILER_ID "GNU")
+set(CMAKE_CXX_COMPILER_ID "Clang")
 
 # The include file search paths:
 set(CMAKE_CXX_TARGET_INCLUDE_PATH
