+++ cmake-build-debug/CMakeFiles/benchmarks.dir/depend.make
@@ -1,20 +1,2 @@
-# CMAKE generated file: DO NOT EDIT!
-# Generated by "Unix Makefiles" Generator, CMake Version 3.17
-
-CMakeFiles/benchmarks.dir/benchmarks/bench.cpp.o: ../benchmarks/bench.cpp
-CMakeFiles/benchmarks.dir/benchmarks/bench.cpp.o: ../benchmarks/includes/3thparty/CLI11.hpp
-CMakeFiles/benchmarks.dir/benchmarks/bench.cpp.o: ../benchmarks/includes/aggregate_tests.h
-CMakeFiles/benchmarks.dir/benchmarks/bench.cpp.o: ../benchmarks/includes/generator.h
-CMakeFiles/benchmarks.dir/benchmarks/bench.cpp.o: ../benchmarks/includes/prepare.h
-CMakeFiles/benchmarks.dir/benchmarks/bench.cpp.o: ../benchmarks/includes/tests.h
-CMakeFiles/benchmarks.dir/benchmarks/bench.cpp.o: ../hashmap_implementations/LPmap.h
-CMakeFiles/benchmarks.dir/benchmarks/bench.cpp.o: ../hashmap_implementations/LPmap2.h
-CMakeFiles/benchmarks.dir/benchmarks/bench.cpp.o: ../hashmap_implementations/helpers.h
-CMakeFiles/benchmarks.dir/benchmarks/bench.cpp.o: ../hashmap_implementations/nodemap.h
-CMakeFiles/benchmarks.dir/benchmarks/bench.cpp.o: ../hashmap_implementations/nodemap1b.h
-CMakeFiles/benchmarks.dir/benchmarks/bench.cpp.o: ../hashmap_implementations/nodemap2.h
-CMakeFiles/benchmarks.dir/benchmarks/bench.cpp.o: ../tools/random.h
-
-CMakeFiles/benchmarks.dir/benchmarks/generator.cpp.o: ../benchmarks/generator.cpp
-CMakeFiles/benchmarks.dir/benchmarks/generator.cpp.o: ../benchmarks/includes/generator.h
-
+# Empty dependencies file for benchmarks.
+# This may be replaced when dependencies are built.
