+++ cmake-build-debug/CMakeFiles/benchmarks.dir/link.txt
@@ -1,1 +1,1 @@
-/usr/bin/g++  -o3  -march=native -g   CMakeFiles/benchmarks.dir/benchmarks/generator.cpp.o CMakeFiles/benchmarks.dir/benchmarks/bench.cpp.o  -o benchmarks 
+/usr/bin/clang++  -o3  -march=native -fsanitize=address -fno-omit-frame-pointer -g   CMakeFiles/benchmarks.dir/benchmarks/generator.cpp.o CMakeFiles/benchmarks.dir/benchmarks/bench.cpp.o  -o benchmarks
