+++ cmake-build-debug/CMakeFiles/clion-log.txt
@@ -1,4 +1,4 @@
-/snap/clion/139/bin/cmake/linux/bin/cmake -DCMAKE_BUILD_TYPE=Debug -DCMAKE_C_COMPILER=/usr/bin/gcc -DCMAKE_CXX_COMPILER=/usr/bin/g++ -G "CodeBlocks - Unix Makefiles" /home/MassiveAtoms/Desktop/PO2
+/snap/clion/139/bin/cmake/linux/bin/cmake -DCMAKE_BUILD_TYPE=Debug -DCMAKE_C_COMPILER=/usr/bin/clang -DCMAKE_CXX_COMPILER=/usr/bin/clang++ -G "CodeBlocks - Unix Makefiles" /home/MassiveAtoms/Desktop/PO2
 -- Configuring done
 -- Generating done
 -- Build files have been written to: /home/MassiveAtoms/Desktop/PO2/cmake-build-debug
