+++ cmake-build-debug/CMakeFiles/cuckoo-impl.dir/flags.make
@@ -1,8 +1,8 @@
 # CMAKE generated file: DO NOT EDIT!
 # Generated by "Unix Makefiles" Generator, CMake Version 3.17
 
-# compile CXX with /usr/bin/g++
-CXX_FLAGS = -o3  -march=native -g   -std=gnu++17
+# compile CXX with /usr/bin/clang++
+CXX_FLAGS = -o3  -march=native -fsanitize=address -fno-omit-frame-pointer -g   -std=gnu++17
 
 CXX_DEFINES =
