+++ cmake-build-debug/Testing/Temporary/LastTest.log
@@ -1,3 +1,3 @@
-Start testing: Mar 05 10:45 -03
+Start testing: Mar 06 13:42 -03
 ----------------------------------------------------------
-End testing: Mar 05 10:45 -03
+End testing: Mar 06 13:42 -03
