+++ cmake-build-debug/compile_commands.json
@@ -1,27 +1,27 @@
 [
 {
   "directory": "/home/MassiveAtoms/Desktop/PO2/cmake-build-debug",
-  "command": "/usr/bin/g++    -o3  -march=native -g   -std=gnu++17 -o CMakeFiles/sagar-test.dir/test/hashmap_test.cpp.o -c /home/MassiveAtoms/Desktop/PO2/test/hashmap_test.cpp",
+  "command": "/usr/bin/clang++    -o3  -march=native -fsanitize=address -fno-omit-frame-pointer -g   -std=gnu++17 -o CMakeFiles/sagar-test.dir/test/hashmap_test.cpp.o -c /home/MassiveAtoms/Desktop/PO2/test/hashmap_test.cpp",
   "file": "/home/MassiveAtoms/Desktop/PO2/test/hashmap_test.cpp"
 },
 {
   "directory": "/home/MassiveAtoms/Desktop/PO2/cmake-build-debug",
-  "command": "/usr/bin/g++    -o3  -march=native -g   -std=gnu++17 -o CMakeFiles/cuckoo-impl.dir/hashmap_implement/cuckoo.cpp.o -c /home/MassiveAtoms/Desktop/PO2/hashmap_implement/cuckoo.cpp",
+  "command": "/usr/bin/clang++    -o3  -march=native -fsanitize=address -fno-omit-frame-pointer -g   -std=gnu++17 -o CMakeFiles/cuckoo-impl.dir/hashmap_implement/cuckoo.cpp.o -c /home/MassiveAtoms/Desktop/PO2/hashmap_implement/cuckoo.cpp",
   "file": "/home/MassiveAtoms/Desktop/PO2/hashmap_implement/cuckoo.cpp"
 },
 {
   "directory": "/home/MassiveAtoms/Desktop/PO2/cmake-build-debug",
-  "command": "/usr/bin/g++    -o3  -march=native -g   -std=gnu++17 -o CMakeFiles/node-build.dir/hashmap_implementations/test.cpp.o -c /home/MassiveAtoms/Desktop/PO2/hashmap_implementations/test.cpp",
+  "command": "/usr/bin/clang++    -o3  -march=native -fsanitize=address -fno-omit-frame-pointer -g   -std=gnu++17 -o CMakeFiles/node-build.dir/hashmap_implementations/test.cpp.o -c /home/MassiveAtoms/Desktop/PO2/hashmap_implementations/test.cpp",
   "file": "/home/MassiveAtoms/Desktop/PO2/hashmap_implementations/test.cpp"
 },
 {
   "directory": "/home/MassiveAtoms/Desktop/PO2/cmake-build-debug",
-  "command": "/usr/bin/g++    -o3  -march=native -g   -std=gnu++17 -o CMakeFiles/benchmarks.dir/benchmarks/generator.cpp.o -c /home/MassiveAtoms/Desktop/PO2/benchmarks/generator.cpp",
+  "command": "/usr/bin/clang++    -o3  -march=native -fsanitize=address -fno-omit-frame-pointer -g   -std=gnu++17 -o CMakeFiles/benchmarks.dir/benchmarks/generator.cpp.o -c /home/MassiveAtoms/Desktop/PO2/benchmarks/generator.cpp",
   "file": "/home/MassiveAtoms/Desktop/PO2/benchmarks/generator.cpp"
 },
 {
   "directory": "/home/MassiveAtoms/Desktop/PO2/cmake-build-debug",
-  "command": "/usr/bin/g++    -o3  -march=native -g   -std=gnu++17 -o CMakeFiles/benchmarks.dir/benchmarks/bench.cpp.o -c /home/MassiveAtoms/Desktop/PO2/benchmarks/bench.cpp",
+  "command": "/usr/bin/clang++    -o3  -march=native -fsanitize=address -fno-omit-frame-pointer -g   -std=gnu++17 -o CMakeFiles/benchmarks.dir/benchmarks/bench.cpp.o -c /home/MassiveAtoms/Desktop/PO2/benchmarks/bench.cpp",
   "file": "/home/MassiveAtoms/Desktop/PO2/benchmarks/bench.cpp"
 }
 ]
