+++ test/hashmap_test.h
